moved ProxmarkVersion to behavior package
